requirement.txt
langchain
langchain-community
openai
faiss-cpu
pypdf
flask
python-dotenv
huggingface-hub
requests
flask
twilio
googlesearch-python
